requirements.txt
requests
pyyaml
schedule
python-dotenv
python-dateutil
flask
gunicorn
mcstatus